add root password

+++ build_all.sh
@@ -91,4 +91,10 @@ else
 fi
 bash ../scripts/pack.sh -t ${TYPE} -v ${VERSION}
 
-mv sdcard.img.xz ubuntu-${VERSION}-${TYPE}-${ARCH}-${BOARD}.img.xz
+if [ -f sdcard.img.xz ];then
+    mv sdcard.img.xz ubuntu-${VERSION}-${TYPE}-${ARCH}-${BOARD}.img.xz
+    echo "build success."
+else
+    echo "sdcard.img.xz not found, build sdcard image failed!"
+    exit 2
+fi
